// src/lib/sumit/types.generated.ts
export interface components {
    schemas: {
// ...
        "OfficeGuy.Apps.Billing.MVC.API.PaymentsController_Payments_Charge_Response": {
            Payment?: null & components["schemas"]["OfficeGuy.Apps.Billing.MVC.API.Typed.Payment"];
            /**
             * Format: int64
             * @description Document ID (OfficeGuy identifier)
             */
            DocumentID?: null | number;
            /**
             * Format: int64
             * @description Document number
             */
            DocumentNumber?: null | number;
            /**
             * Format: int64
             * @description Customer number
             */
            CustomerID?: null | number;
            /**
             * @description Document download URL
             *     Produced document will be original (on first fetching), or certified copy.
             */
            DocumentDownloadURL?: null | string;
        };
// ...
        "OfficeGuy.Apps.Billing.MVC.API.Typed.Payment": {
            /**
             * Format: int64
             * @description Payment identifier
             */
            ID?: number;
            /**
             * Format: int64
             * @description Customer identifier
             */
            CustomerID?: number;
            /**
             * Format: date-time
             * @description Payment date
             */
            Date?: null | string;
            /** @description Indicates if the payment is valid */
            ValidPayment?: boolean;
            /** @description Payment status */
            Status?: null | string;
            /** @description Payment status description */
            StatusDescription?: null | string;
            /**
             * Format: double
             * @description Payment amount
             */
            Amount?: number;
            Currency?: null & components["schemas"]["Accounting_Typed_DocumentCurrency"];
            /** @description Payment method details */
            PaymentMethod?: null & components["schemas"]["OfficeGuy.Apps.Billing.MVC.API.Typed.PaymentMethod"];
            /** @description Authorization number */
            AuthNumber?: null | string;
            /**
             * Format: double
             * @description First credit card installments payment amount
             */
            FirstPaymentAmount?: null | number;
            /**
             * Format: double
             * @description Non first credit card installments payment amount
             */
            NonFirstPaymentAmount?: null | number;
            /** @description Relevant only for payments originating from recurring payments */
            RecurringCustomerItemIDs?: null | number[];
        };
        "OfficeGuy.Apps.Billing.MVC.API.Typed.PaymentMethod": {
            /** Format: int64 */
            ID?: null | number;
            /** Format: int64 */
            CustomerID?: null | number;
            /**
             * @description Credit card full number
             *     Required for credit card
             */
            CreditCard_Number?: null | string;
            /**
             * @description Credit card last 4 digits
             *     Shouldn't be input by API caller
             */
            CreditCard_LastDigits?: null | string;
            /**
             * Format: int32
             * @description Credit card expiration month (1-2 digits)
             *     Required for credit card
             */
            CreditCard_ExpirationMonth?: null | number;
            /**
             * Format: int32
             * @description Credit card expiration year (4 digits)
             *     Required for credit card
             */
            CreditCard_ExpirationYear?: null | number;
            /**
             * @description Credit card CVV/CVV2
             *     Required when CVV is required by credit company
             */
            CreditCard_CVV?: null | string;
            /** @description Credit card Track2 */
            CreditCard_Track2?: null | string;
            /**
             * @description Credit card owner Israel Citizen ID / Passport Number
             *     Required when Citizen ID is required by credit company
             */
            CreditCard_CitizenID?: null | string;
            /**
             * @description Credit card mask
             *     Shouldn't be input by API caller
             */
            CreditCard_CardMask?: null | string;
            /** @description Credit card token */
            CreditCard_Token?: null | string;
            /**
             * Format: int64
             * @description Direct debit bank number
             *     For instance, 12 indicates Bank Hapoalim.
             */
            DirectDebit_Bank?: null | number;
            /**
             * Format: int64
             * @description Direct debit bank branch number
             */
            DirectDebit_Branch?: null | number;
            /**
             * Format: int64
             * @description Direct debit bank account number
             */
            DirectDebit_Account?: null | number;
            /**
             * Format: date-time
             * @description Direct debit expiration date
             */
            DirectDebit_ExpirationDate?: null | string;
            /**
             * Format: int64
             * @description Direct debit maximum charge amount
             */
            DirectDebit_MaximumAmount?: null | number;
            /** @description Payment method type */
            Type?: components["schemas"]["OfficeGuy.Apps.Billing.MVC.API.Typed.PaymentMethodType"];
        };
// ...
        "Response_OfficeGuy.Apps.Billing.MVC.API.PaymentsController_Payments_Charge_Response": {
            /** @description Response status */
            Status?: components["schemas"]["Teva.Common.ResponseStatus"];
            /** @description Error message, in a user readable format */
            UserErrorMessage?: null | string;
            /** @description Technical error details, let us know if you received this. */
            TechnicalErrorDetails?: null | string;
            /** @description API specific response data */
            Data?: null & components["schemas"]["OfficeGuy.Apps.Billing.MVC.API.PaymentsController_Payments_Charge_Response"];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}